pyproject.toml
[project]
name = "unidb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.114.1",
    "password-validator>=1.0",
    "prisma>=0.15.0",
    "python-dotenv>=1.0.1",
    "supabase>=2.9.0",
    "textual>=0.82.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.9",
    "textual-dev>=1.6.1",
]
